Role chua test

+++ WebAPI/Controllers/ExpertController.cs
@@ -66,12 +66,12 @@ namespace WebAPI.Controllers
         {
             try
             {
-                //var role = User.Claims.FirstOrDefault(c => c.Type == ClaimTypes.Role)?.Value;
-                //if (!string.IsNullOrEmpty(role))
-                //{
-                //if (role == "ADMIN" && role == "STAFF")
-                //{
-                var expert = Validate(expertVM);
+                var role = User.Claims.FirstOrDefault(c => c.Type == ClaimTypes.Role)?.Value;
+                if (!string.IsNullOrEmpty(role))
+                {
+                    if (role == CommonValues.ADMIN || role == CommonValues.STAFF)
+                    {
+                        var expert = Validate(expertVM);
                 var check = expertService.Add(expert);
                 return await check ? Ok(new
                 {
@@ -82,20 +82,20 @@ namespace WebAPI.Controllers
                     Status = 0,
                     Message = "Add Expert Fail"
                 });
-                //}
-                //else
-                // return   //Ok(new
-                //{
-                //    Status = 0,
-                //    Message = "Role Denied",
-                //    Data = new { }
-                //});
-
-                //else
-                //{
-                //    return Unauthorized();
-                //}
             }
+                else
+                return Ok(new
+                {
+                    Status = 0,
+                    Message = "Role Denied",
+                    Data = new { }
+                });
+        }
+                else
+                {
+                    return Unauthorized();
+                 }
+                }
             catch (Exception ex)
             {
                 return BadRequest(ex.Message);
@@ -112,12 +112,12 @@ namespace WebAPI.Controllers
 
             try
             {
-                //var role = User.Claims.FirstOrDefault(c => c.Type == ClaimTypes.Role)?.Value;
-                //if (!string.IsNullOrEmpty(role))
-                //{
-                //    if (role == "ADMIN" && role == "STAFF")
-                //    {
-                var expert = Validate(expertVM);
+                var role = User.Claims.FirstOrDefault(c => c.Type == ClaimTypes.Role)?.Value;
+                if (!string.IsNullOrEmpty(role))
+                {
+                    if (role == CommonValues.ADMIN || role == CommonValues.STAFF)
+                    {
+                        var expert = Validate(expertVM);
                 var check = expertService.Update(expert);
                 return await check ? Ok(new
                 {
@@ -128,21 +128,21 @@ namespace WebAPI.Controllers
                     Status = 0,
                     Message = "Update Expert Fail"
                 });
-                //}
-                //    else
-                //    {
-                //        return Ok(new
-                //        {
-                //            Status = 0,
-                //            Message = "Role Denied",
-                //            Data = new { }
-                //        });
-                //    }
-                //}
-                //else
-                //{
-                //    return Unauthorized();
-                //}
+                    }
+                    else
+                    {
+                        return Ok(new
+                        {
+                            Status = 0,
+                            Message = "Role Denied",
+                            Data = new { }
+                        });
+                    }
+                }
+                else
+                {
+                    return Unauthorized();
+                }
             }
             catch (Exception ex)
             {
@@ -161,36 +161,37 @@ namespace WebAPI.Controllers
 
             try
             {
-                //var role = User.Claims.FirstOrDefault(c => c.Type == ClaimTypes.Role)?.Value;
-                //if (!string.IsNullOrEmpty(role))
-                //{
-                //if (role == "ADMIN" || role == "STAFF")
-                //{
-                var check = expertService.Delete(id);
-                return await check ? Ok(new
+                var role = User.Claims.FirstOrDefault(c => c.Type == ClaimTypes.Role)?.Value;
+                if (!string.IsNullOrEmpty(role))
                 {
-                    Status = 1,
-                    Message = "Delete Expert Success"
-                }) : Ok(new
+                    if (role == CommonValues.ADMIN || role == CommonValues.STAFF)
+                    {
+                        var check = expertService.Delete(id);
+                        return await check ? Ok(new
+                        {
+                            Status = 1,
+                            Message = "Delete Expert Success"
+                        }) : Ok(new
+                        {
+                            Status = 0,
+                            Message = "Delete Expert Fail"
+                        });
+                    }
+                    else
+                    {
+                        return Ok(new
+                        {
+                            Status = 0,
+                            Message = "Role Denied",
+                            Data = new { }
+                        });
+                    }
+                }
+                else
                 {
-                    Status = 0,
-                    Message = "Delete Expert Fail"
-                });
-                //}
-                //else
-                //{
-                //    return Ok(new
-                //    {
-                //        Status = 0,
-                //        Message = "Role Denied",
-                //        Data = new { }
-                //    });
-                //}
-                //else
-                //{
-                //    return BadRequest();
-                //}
-            }
+                    return BadRequest();
+                }
+                }
             catch (Exception ex)
             {
                 return BadRequest(ex.Message);
@@ -289,6 +290,8 @@ namespace WebAPI.Controllers
 
             return new JwtSecurityTokenHandler().WriteToken(token);
         }
+
+        
 
         [HttpPost]
         public async Task<IActionResult> LoginExpert(ExpLogin expLogin)
